Move pre-commit hooks to GitHub Actions (#487)
We currently use Husky to run unit tests and linting as pre-commit hooks. This means that we cannot commit often, as we have to wait for the tests to run. We are okay with broken/unlinted code being committed and pushed in a branch as long as this doesn't get merged! We therefore remove the pre-commit hooks and update our CI to test and lint.