SDDSIP-1101: Fix 500 error when going back after designated site removed
https://eaflood.atlassian.net/browse/SDDSIP-1101

Going through the "Give details of conservation issues" flow, we enter details of a designated site. At the end, we get a "Check your answers" page with a "Remove site" option. Removing the site and then using the Back link (or browser back button) gives us an error when we hit a page such as `/advice-from-natural-england`.