Fix: db migrations `no pg_hba.conf entry` error (#626)
Our db migrations aren't running in the dev environment, giving the following error:

```
no pg_hba.conf entry for host "...", user "...", database "...", no encryption
```

It seems this may be due to the db expecting an SSL connection. We therefore update the `sequelize.config.cjs` file we use when running migrations to set the required options.

Since we are applying the exact same configuration in every environment (with the only changes being down to the content of the environment variables) we also refactor the file to apply the same configuration object to each environment.

Note that as per `packages/connectors-lib/src/sequelize.js`, the SSL config is conditional based on the content of the `POSTGRES_NOSSL` env var; if this is set to `true` then we don't apply the SSL config. It may be required to set `POSTGRES_NOSSL=true` when running migrations locally.

The migration initially failed in the dev environment as the columns it tries to create already exist. Sequelize doesn't give an easy way of forcing migrations to be marked as done, so instead we log into the db and manually add the migration to the table of run migrations:

```sql
INSERT INTO "SequelizeMeta" (name) VALUES ('20240110142636-migration.cjs');
```