fix: Empty address drop-down in site address page (#633)
https://eaflood.atlassian.net/browse/SDDSIP-2084

The "What is the site address?" page in the "Give site information" flow didn't seem to have been fully updated for the v2 address lookup API, as it was still expecting to find address data in the older format (ie. each address contains `Address.UPRN` and `Address.AddressLine`, instead of just `uprn` and `addressLine`). This led to the page's address down-down being empty. We therefore update the template to extract the data correctly.