Fix bug in fetch-helper
It turns out that using `+` to concatenate strings doesn't play well with the ternary operator; something like `const str = 'abc' + condition ? 'd' : 'z'` will always result in `abcd` regardless of whether or not `condition` is truthy. This means that with a certain type of error, we will always try to call `err.response.json()` regardless of whether or not its content is json. We therefore fix this.