fix: Improve logging in `fetch-helper` (#650)
Currently, when the `connectors-lib:fetch` debug option is set and we make a request to the address lookup service, the request options add ~3500 lines to the debug logs.

The problem is effectively twofold:

- We don't want to output the entire contents of `agent.options.key.data` and `agent.options.cert.data`;
- Even with those removed, the object is still spread across multiple lines in the logs.

We therefore add a replacer function we can use when stringifying the object which will redact the long buffers. We also ensure the output is all on one line; this means that it's no longer spread across multiple lines in the logs, but if we expand the line in CloudWatch we will still see a nicely-formatted JSON object.

While fixing SonarCloud errors for the PR, we realised there was no test coverage, so we put some quick tests in place so we could be sure that our fixes don't break anything. In the process we fixed a bug in the logging:

```js
   if (err.response) {
      const msg = 'response error: ' + err.response.headers.get('content-type').includes(APPLICATION_JSON)
        ? JSON.stringify(await err.response.json())
        : await err.response.body()
      console.error(`Unknown error thrown in fetch: ${msg}`)
      throw new Error(msg)
    }
```

This looks like it should be fine, however it always tries to call `err.response.json()`. This is because it evaluates `'response error: ' + err.response.headers.get('content-type').includes(APPLICATION_JSON)` first, giving a string `response error: true` or `response error: false`. It then uses this for the condition of the ternary, and because the resulting string is always truthy, we therefore always end up trying to perform `JSON.stringify(await err.response.json())`

We fix this by refactoring to use string templates as we do elsewhere in the file.